Navigation component was added
