+++ src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import './App.css';
-import Welcome from './components/Welcome/Welcome';
-import Header from './components/Header/Header';
-import Article from './components/Article/Article';
+import Welcome from './components/Welcome';
+import Header from './components/Header';
+import Article from './components/Article';
 // Move each BEM block to a separate component (file) and import them here
 
 function App() {
