+++ src/components/Header/index.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Navigation from '../Navigation';
+import './style.css';
+
+const Header = () => (
+  <header className="header">
+    <h1 className="header__title">Site Name</h1>
+    <Navigation />
+  </header>
+);
+
+export default Header;
