+++ src/components/Header/style.css
@@ -18,21 +18,6 @@
   font-size: inherit;
 }
 
-.navigation__link {
-  display: inline-block;
-  outline: none;
-  text-decoration: none;
-  opacity: 0.7;
-  padding: 0 0.5em;
-  color: black;
-  transition: opacity 0.2s ease-in-out;
-}
-
-.navigation__link:hover,
-.navigation__link:focus {
-  opacity: 1;
-}
-
 @media (min-width: 500px) {
   .header {
     text-align: left;
