+++ src/components/Navigation/index.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import './style.css';
+
+const Navigation = () => (
+  <nav className="navigation">
+    <a className="navigation__link" href="#about">
+      About
+    </a>
+    <a className="navigation__link" href="#services">
+      Services
+    </a>
+    <a className="navigation__link" href="#contact">
+      Contact
+    </a>
+  </nav>
+);
+
+export default Navigation;
