+++ src/components/Welcome/index.jsx
@@ -5,7 +5,7 @@
 
 // Add a default export statement for Welcome component to use it in the other files
 import React from 'react';
-import './Welcome.css';
+import './style.css';
 
 const Welcome = () => (
   <section className="welcome">
